+++ src/main/java/shapes/exceptions/NotFoundException.java
@@ -1,7 +1,7 @@
 package shapes.exceptions;
 
 public class NotFoundException extends Throwable {
-    public NotFoundException(int id, Class clazz) {
+    public NotFoundException(long id, Class clazz) {
         super(String.format("%s with id = %s is not found", clazz.getSimpleName(), id));
     }
 }
