+++ src/main/java/shapes/repositories/RadiusInfoRepository.java
@@ -3,5 +3,5 @@ package shapes.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shapes.models.RadiusInfo;
 
-public interface RadiusInfoRepository extends JpaRepository<RadiusInfo, Integer> {
+public interface RadiusInfoRepository extends JpaRepository<RadiusInfo, Long> {
 }
